pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-indices-bot"
version = "0.1.0"
description = "Deriv synthetic indices mean-reversion bot with research harness"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "websockets>=12.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sidx*"]
